chore: remove duplicated impls
