+++ lib/api/src/grpc/proto/qdrant.proto
@@ -8,7 +8,6 @@ import "points_internal_service.proto";
 import "qdrant_internal_service.proto";
 import "raft_service.proto";
 import "shard_snapshots_service.proto";
-import "shard_storage_read_service.proto";
 import "snapshots_service.proto";
 import "storage_read_service.proto";
