+++ src/tonic/api/mod.rs
@@ -4,14 +4,12 @@ pub mod points_api;
 pub mod points_internal_api;
 pub mod qdrant_internal_api;
 pub mod raft_api;
-pub mod shard_storage_read_api;
 pub mod snapshots_api;
 pub mod storage_read_api;
 pub mod telemetry_wrapper;
 
 mod collections_common;
 mod query_common;
-mod storage_read_common;
 mod update_common;
 
 use collection::operations::validation;
