+++ src/tonic/mod.rs
@@ -22,7 +22,6 @@ use ::api::grpc::qdrant::points_server::PointsServer;
 use ::api::grpc::qdrant::qdrant_internal_server::QdrantInternalServer;
 use ::api::grpc::qdrant::qdrant_server::{Qdrant, QdrantServer};
 use ::api::grpc::qdrant::shard_snapshots_server::ShardSnapshotsServer;
-use ::api::grpc::qdrant::shard_storage_read_server::ShardStorageReadServer;
 use ::api::grpc::qdrant::snapshots_server::SnapshotsServer;
 use ::api::grpc::qdrant::storage_read_server::StorageReadServer;
 use ::api::grpc::qdrant::{HealthCheckReply, HealthCheckRequest};
@@ -49,7 +48,6 @@ use crate::tonic::api::collections_internal_api::CollectionsInternalService;
 use crate::tonic::api::points_api::PointsService;
 use crate::tonic::api::points_internal_api::PointsInternalService;
 use crate::tonic::api::qdrant_internal_api::QdrantInternalService;
-use crate::tonic::api::shard_storage_read_api::ShardStorageReadService;
 use crate::tonic::api::snapshots_api::{ShardSnapshotsService, SnapshotsService};
 use crate::tonic::api::storage_read_api::StorageReadService;
 use crate::tonic::api::telemetry_wrapper::{
@@ -266,8 +264,6 @@ pub fn init_internal(
                 QdrantInternalService::new(telemetry_collector, settings, consensus_state.clone());
             let collections_internal_service = CollectionsInternalService::new(toc.clone());
             let shard_snapshots_service = ShardSnapshotsService::new(toc.clone(), http_client);
-            let shard_storage_read_service =
-                ShardStorageReadService::<StorageBackend>::new(toc.clone());
             let raft_service =
                 RaftService::new(to_consensus, consensus_state, tls_config.is_some());
 
@@ -333,12 +329,6 @@ pub fn init_internal(
                     .send_compressed(CompressionEncoding::Gzip)
                     .accept_compressed(CompressionEncoding::Gzip)
                     .max_decoding_message_size(usize::MAX),
-                )
-                .add_service(
-                    ShardStorageReadServer::new(shard_storage_read_service)
-                        .send_compressed(CompressionEncoding::Gzip)
-                        .accept_compressed(CompressionEncoding::Gzip)
-                        .max_decoding_message_size(usize::MAX),
                 )
                 .add_service(
                     RaftServer::new(raft_service)
